+++ babel.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  presets: [['@babel/preset-env', { targets: { esmodules: true } }]],
+  plugins: ['@babel/plugin-transform-typescript'],
+};