+++ jest.config.js
@@ -1,4 +1,6 @@
 module.exports = {
-  preset: 'ts-jest',
+  transform: {
+    '^.+\\.[t|j]sx?$': 'babel-jest',
+  },
   testEnvironment: 'jsdom',
 };